Add files via upload
Designed a smart electric two-wheelerfor online bike rental service which has a
mobile web-page accessed through a QR code forthe users. These vehicles are
one of the bestreplacement forthe traditional public transportation system. It
can reduce the waiting time up to 90% and rushing of passengers in the public
transport up to 30%. It was achieved by utilising CellularIoT tech integrated in
the embedded system of the vehicle.